Move grouping functions to modules.
